extract swapchain_present
